Moved debounce to helpers removed delay from index.js
