--- src/css-units-to-pixels/css-fixed-units.js
@@ -1,38 +1,0 @@
-/**
- * CSS Fixed units with pixel values.
- */
-const CSSFixedUnits = [{
-    unit: 'xx-small',
-    // 9 / 16
-    PXValue: 9,
-}, {
-    unit: 'x-small',
-    // 10 / 16
-    PXValue: 10,
-}, {
-    unit: 'small',
-    // 13 / 16
-    PXValue: 13,
-}, {
-    unit: 'normal',
-    // 16 / 16
-    PXValue: 16,
-}, {
-    unit: 'medium',
-    // 16 / 16
-    PXValue: 16,
-}, {
-    unit: 'large',
-    // 18 / 16
-    PXValue: 18,
-}, {
-    unit: 'x-large',
-    // 24 / 16
-    PXValue: 24,
-}, {
-    unit: 'xx-large',
-    // 32 / 16
-    PXValue: 32,
-}];
-
-export default CSSFixedUnits;
