--- src/css-units-to-pixels/css-units.js
@@ -1,62 +1,0 @@
-/**
- * CSS Units with pixel factor.
- */
-const CSSUnits = [{
-    unit: 'px',
-    //  1 / 16
-    PXFactor: 1,
-}, {
-    unit: 'pt',
-    // 96 / 72 / 16
-    PXFactor: 1.33333333328,
-}, {
-    unit: 'em',
-    // 1
-    PXFactor: 16,
-}, {
-    unit: 'rem',
-    // 1
-    PXFactor: 16,
-}, {
-    unit: 'cm',
-    // ???
-    PXFactor: 16,
-}, {
-    unit: 'q',
-    // ???
-    PXFactor: 16,
-}, {
-    unit: 'in',
-    // ???
-    PXFactor: 16,
-}, {
-    unit: 'pc',
-    // 96 / 72 / 12 / 16
-    PXFactor: 0.11111111104,
-}, {
-    unit: 'ex',
-    // ???
-    PXFactor: 16,
-}, {
-    unit: 'ch',
-    // ???
-    PXFactor: 16,
-}, {
-    unit: 'vw',
-    // Needs override.
-    PXFactor: 16,
-}, {
-    unit: 'vh',
-    // Needs override.
-    PXFactor: 16,
-}, {
-    unit: 'vmin',
-    // Needs override.
-    PXFactor: 16,
-}, {
-    unit: 'vmax',
-    // Needs override.
-    PXFactor: 16,
-}];
-
-export default CSSUnits;
