--- src/store.js
@@ -1,4 +1,0 @@
-export default {
-	rootFontSize: NaN,
-	resizeSettings: null
-}
